Almost done the clean command.
I had to implement some list functions, and while I was at it, I also
implemented the "list" command, which calls those functions.

Instead of having library functions to clean the packages and cache, I
figured it was better/safer to rely on the code in "cmd/pm" to figure out
what to do with the various results of the List* functions.

references #2